pyproject.toml
[tool.poetry]
name = "flicks-api"
version = "0.1.0"
description = ""
authors = ["Isaac Adewumi <ienioladewumi@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
django-environ = "^0.11.2"
huey = "^2.4.5"
whitenoise = "^6.5.0"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
django-cors-headers = "^4.2.0"
redis = "^4.6.0"
psycopg2-binary = "^2.9.7"
django-extra-checks = "^0.13.3"
django-fernet-encrypted-fields = "^0.1.3"
django = "^4.2.5"
drf-yasg = "^1.21.7"
setuptools = "^68.2.2"
solana = "^0.30.2"
pynacl = "^1.5.0"
requests = "^2.31.0"
aws-sns-message-validator = "^0.0.5"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.289"
black = "^23.9.1"
isort = "^5.12.0"
pylint = "^2.17.5"
ipython = "^8.15.0"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
skip-string-normalization = true
target-version = ['py39']
exclude = '''
  /(
    | \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
'''

[tool.ruff]
preview = true
line-length = 119
select = [
    "E",
    "F",
    "W",
    "C90",
    "C4",
    "N",
    "B",
    "G",
    "DJ",
    "ISC",
    "ICN",
    "T10",
    "T20",
    "UP",
    "YTT",
    "S",
    "BLE",
    "FBT",
    "A",
    "Q",
    "ERA",
    "PGH",
    "RET",
    "SIM",
    "TID",
    "ARG",
    "PL",
    "PIE",
    "PYI",
    "COM",
    "DTZ",
    "EXE",
    "INP",
    "TCH",
    "PTH",
    "INT",
    "TRY",
    "RSE",
    "SLF",
    "RUF",
    "FURB",
    "PERF",
    "FLY",
    "TD",
]
exclude = ["migrations", ".venv"]
ignore = ["TRY003", "G004", "TRY301"]

[tool.ruff.per-file-ignores]
"**/test_**.py" = ["S101"]
"conf/settings.py" = ["E402"]

[tool.ruff.flake8-quotes]
inline-quotes = "single"
docstring-quotes = "double"
multiline-quotes = "double"

[tool.ruff.pylint]
max-args = 6

[tool.isort]
profile = "black"
line_length = 119
length_sort = true
force_grid_wrap = 0
multi_line_output = 3
use_parentheses = true
known_django = "django"
combine_as_imports = true
lines_between_sections = 1
include_trailing_comma = true
force_sort_within_sections = true
known_rest_framework = "rest_framework"
known_apps = "apps"
known_services = "services"
known_utils = "utils"
skip_glob = ["**/.venv/*", "**/migrations/*"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY","DJANGO", "REST_FRAMEWORK", "FIRSTPARTY", "UTILS", "SERVICES", "APPS", "LOCALFOLDER"]

